Reinstate text renderer changes lost in previous merge
The changes are done in a somewhat different way to fit in with the
LimitPreviousContext changes. A couple of extra rectangles are
figured out, which we could avoid, but at the cost of a LOT more
conditional compilation. I doubt it is worth it. If the USETEXTRENDERER
stuff turns out to be a good idea, we can get rid of the brush
and tweak the arguments to the drawString Action to match the new approach.
Then we can get rid of GetTextRendererBoundsAndAlignment and just
figure them out directly in the code that calls drawString and
figures the offset adjustment.

Also cleans up: there is no longer a reason to pass the brush argument
to drawString.