Work around and gather info on HT-9
Brought in a couple classes from Bloom that we have found helpful in dealing with these hard-to-reproduce errors. They allow us to show a "toast" that can be ignored or clicked on to send a report.

Also retrying getting brushes a couple times in case this is a timing bug. Also fall back to an all-green trackbar if even that fails.