HT-188: Changed appcast.xml for 1.5 branch to look for the latest version in the 1.5 and the code to look for the appcast.xml file from the latest successful 1.5 build.
HT-190: Updated the link URL in appcast.XML to point to software.sil.org/hearthis/download
HT-191: Updated release notes to indicate support for Paratext 8 projects and distinguish between HT 1.4.x and 1.5.x.
HT-192: Updated the enclosure URL (to access the latest msi) in appcast.XML to include the /r folder so we don't get a 404.