Update Toolbar Buttons
All buttons are now 16px high. All now use a simple custom control and ones with state (checked or not) now show a rectangle when 'checked'. All use only a single color, the "CommonMuted" one, so they work across color themes. The actual interactions with the mouse could be improved.