Removed references to Spart and Paratext.LexicalContracts. These DLLs are part of the
ParatextData NuGet package, but they are not needed in HearThis and have never been included
in the Installer. Also removed reference to Microsoft.DotNet.PlatformAbstractions, which is
supposedly needed by Microsoft.Extensions.DependencyModel. It's supposed to help with
testability, but the unit tests don't need it, and it has also never been included in the Installer.