Changed NuGet.targets to use a slash instead of a space following the SolutionDir
This allows it to work with the latest version of NuGet.
Since some packages that HearThis needs require a more recent version of NuGet, I also added a conditional command to update nuget to the latest version if it already exists.
(See https://github.com/NuGet/Home/issues/7657 for background)