HT-74 [WIP] Added some explanatory text to the Shift Clips dialog
Also, threw in a conceptual design idea of allowing the user to determine the direction in the dialog by means of radio buttons. (This is not implemented and is not a coherent design, since the current logic when the menu item is chosen does its check based on right-shifting only.