Added bootstrapper project to install VC++ redistributable
Remove all x86 stuff from installer
Made installer capable of handling debug/release distinction.
WIP/TODO: Need to get a version of the logo (with transparency) that displays in the correct size in the setup dialog. (ico file has various sizes, but by default it shows up tiny and there doesn't seem to be a way to select a larger one.