HT-407, HT-148 [WIP]: Completed most of the UI work and underlying code changes to allow for setting and storing whitespace characters as block breaking and clause breaking characters.
I didn't do any actual testing to see how well it works. I also noted that the CheckComboBox that I got from a nuget package is probably not adequate (in several ways), so we'll either need to make/request some changes or roll our own version of it (if this is the UI approach we want to take).