HT-359: When restoring a clip from backup, set Text back to Original.
Fixed bugs that cause the wrong radio button to get selected as a side-effect of event handlers when user selects an option.
Prevented catch of an exception that is a real bug (fix for that bug forthcoming)
Improved logic for handling state changes in the non-problematic scenarios.
Moved line of code to make the info.xml file get written when an existing clip without recording info is noted as being okay.