[WIP] This was an attempt to reduce the amount of code we have to maintain in HearThis by using the Diffplex library to compute string differences. I had/have
some hope that this approach might also result in somewhat more useful diffs. But there were questions about how well Diffplex would handle data with surrogate pairs and the additional code to handle rejoining words, and making sure diacritics stayed with their base characters was proving non-trivial. Then I found what appears to be a fairly major bug/limitation, and I decided to abandon this. If that bug is fixed, it might be worth considering this approach in the future.