Fix occasional wrong local IP address
Contains the same improvement as is being submitted to Bloom.
The network stack sometimes chooses the network interface and thus HearThis declares the wrong IP address for itself, leading to connect issues with HearThisAndroid. In this change the PC does its own survey of network interfaces and chooses the one with the lowest "interface metric."
This still has some debug output, which I will remove before PR.